Seperate box-creator and new-td files
