+++ lib/index.js
@@ -1,6 +1,8 @@
 require('./styles');
 require('../css/main');
 require('../css/reset');
+let NewTD = require('./new-td');
+let newTDBoxCreator = require('./box-creator');
 
 const $ = require('jquery');
 
@@ -15,28 +17,6 @@ let $task = $('#task-input');
 let $userSearch = $('#search-box');
 let $h2 = $('h2');
 let $p = $('p');
-
-function NewTD(id, title, task, quality) {
-  this.id = id;
-  this.title = title;
-  this.task = task;
-  this.quality = quality;
-}
-
-function newTDBoxCreator(object) {
-  $('.td-container').prepend(`<article id=${object.id} class='td-box'>
-      <div class='flexer'>
-        <h2 class='td-title' contenteditable='true'>${object.title}</h2>
-        <button type='button' name='button' class='delete-button'></button>
-      </div>
-      <p class='td-task' contenteditable='true'>${object.task}</p>
-      <div class='quality-container'>
-        <button type='button' name='button' class='up-button'></button>
-        <button type='button' name='button' class='down-button'></button>
-        <h4 class='quality-rating'>${object.quality}</h4>
-      </div>
-  </article>`);
-}
 
 function setTDStorage(id, object) {
   localStorage.setItem(id, JSON.stringify(object));
