+++ lib/new-td.js
@@ -1,0 +1,8 @@
+function NewTD(id, title, task, quality) {
+  this.id = id;
+  this.title = title;
+  this.task = task;
+  this.quality = quality;
+}
+
+module.exports = NewTD;
